New upstream version 20200619+dfsg.1

+++ recvattach.h
@@ -24,6 +24,7 @@
 #ifndef MUTT_RECVATTACH_H
 #define MUTT_RECVATTACH_H
 
+#include <stdint.h>
 #include <stddef.h>
 #include <stdbool.h>
 #include <stdio.h>
@@ -44,7 +45,7 @@ extern char *C_MessageFormat;
 void mutt_attach_init(struct AttachCtx *actx);
 void mutt_update_tree(struct AttachCtx *actx);
 
-const char *attach_format_str(char *buf, size_t buflen, size_t col, int cols, char op, const char *src, const char *prec, const char *if_str, const char *else_str, unsigned long data, MuttFormatFlags flags);
+const char *attach_format_str(char *buf, size_t buflen, size_t col, int cols, char op, const char *src, const char *prec, const char *if_str, const char *else_str, intptr_t data, MuttFormatFlags flags);
 void mutt_view_attachments(struct Email *e);
 
 void mutt_generate_recvattach_list(struct AttachCtx *actx, struct Email *e, struct Body *parts, FILE *fp, int parent_type, int level, bool decrypted);
